actually make GS user drum work https://github.com/spessasus/spessasynth_core/commit/f9dedba227e7c7a1c80c43e7dc6f3980a7e43cbc

+++ SpessaSharp/Synthesizer/SpessaSharpProcessor.cs
@@ -331,7 +331,7 @@ public sealed class SpessaSharpProcessor
     /// <returns>is an array of voices.</returns>
     internal Synthesizer.CachedVoiceList GetVoicesForPreset(
         SynthPatch preset, int midiNote, int velocity) =>
-        _synthCore.GetVoicesForPreset(preset, midiNote, velocity);
+        _synthCore.GetVoicesForPreset(preset, (byte)midiNote, (byte)velocity);
 
     // Private methods
